+++ front/right_sidebar.py
@@ -1,0 +1,6 @@
+import streamlit as st
+
+def render_right_sidebar():
+    """오른쪽 사이드바 구현"""
+    st.header("Alternative Suggestions")
+    st.write("This section is reserved for future features!")
